User auth

+++ backend/routes/certifications.js
@@ -4,6 +4,8 @@ const router = express.Router()
 const { v4: uuidv4 } = require('uuid')
 const db = require('../utils/db')
 const { sendSuccess, sendError, safeTrim, isValidLength, objFieldMaxLengths } = require('../utils/responses')
+//centralizes user scoping while auth and guest flows are future work.
+const { getCurrentUserID } = require('../utils/users')
 
 // POST new certifications!
 router.post('/', (req,res,next) => {
@@ -16,9 +18,10 @@ router.post('/', (req,res,next) => {
         return sendError(res, 400, "Name, issuer, and date earned are required", { code: "VALIDATION_ERROR", details: {} })
     }
 
-    const strQuery = `INSERT INTO tblCertifications (CertID, Name, Issuer, DateEarned) VALUES (?, ?, ?, ?)`
+    const intUserID = getCurrentUserID(req)
+    const strQuery = `INSERT INTO tblCertifications (CertID, Name, Issuer, DateEarned, UserID) VALUES (?, ?, ?, ?, ?)`
 
-    db.run(strQuery, [strCertID, strName, strIssuer, strDateEarned], (err) => {
+    db.run(strQuery, [strCertID, strName, strIssuer, strDateEarned, intUserID], (err) => {
         if (err) {
             console.error("Error creating certification:", err.message)
             return sendError(res, 500, "Failed to get certifications", { code: "SERVER_ERROR", details: {} })
@@ -30,9 +33,10 @@ router.post('/', (req,res,next) => {
 
 // GET all da certifications
 router.get('/', (req,res,next) => {
-    const strQuery = `SELECT * FROM tblCertifications ORDER BY DateEarned DESC`
+    const intUserID = getCurrentUserID(req)
+    const strQuery = `SELECT * FROM tblCertifications WHERE UserID=? ORDER BY DateEarned DESC`
 
-    db.all(strQuery, [], (err, arrRows) => {
+    db.all(strQuery, [intUserID], (err, arrRows) => {
         if (err) {
             console.error("Error fetching certifications:", err.message)
             return sendError(res, 500, "Failed to get certifications", { code: "SERVER_ERROR" })
@@ -44,9 +48,10 @@ router.get('/', (req,res,next) => {
 // DELETE certification
 router.delete('/:id', (req,res,next) => {
     const strCertID = req.params.id
-    const strQuery = `DELETE FROM tblCertifications WHERE CertID=?`
+    const intUserID = getCurrentUserID(req)
+    const strQuery = `DELETE FROM tblCertifications WHERE CertID=? AND UserID=?`
 
-    db.run(strQuery, [strCertID], function (err) {
+    db.run(strQuery, [strCertID, intUserID], function (err) {
         if (err) {
             console.error("Error deleting certification:", err.message)
             return sendError(res, 500, "Failed to delete certification", { code: "SERVER_ERROR", details: {} })
